solve the problems arising from review
