+++ rocketmq-spring-boot-samples/rocketmq-produce-demo/src/main/java/org/apache/rocketmq/samples/springboot/domain/User.java
@@ -15,35 +15,35 @@
  * limitations under the License.
  */
 
-package org.apache.rocketmq.samples.springboot.consumer;
+package org.apache.rocketmq.samples.springboot.domain;
 
 public class User {
-    private String userName;
-    private Byte userAge;
+        private String userName;
+        private Byte userAge;
 
-    public String getUserName() {
-        return userName;
-    }
+        public String getUserName() {
+            return userName;
+        }
 
-    public User setUserName(String userName) {
-        this.userName = userName;
-        return this;
-    }
+        public User setUserName(String userName) {
+            this.userName = userName;
+            return this;
+        }
 
-    public Byte getUserAge() {
-        return userAge;
-    }
+        public Byte getUserAge() {
+            return userAge;
+        }
 
-    public User setUserAge(Byte userAge) {
-        this.userAge = userAge;
-        return this;
-    }
+        public User setUserAge(Byte userAge) {
+            this.userAge = userAge;
+            return this;
+        }
 
-    @Override
-    public String toString() {
-        return "User{" +
-            "userName='" + userName + '\'' +
-            ", userAge=" + userAge +
-            '}';
+        @Override
+        public String toString() {
+            return "User{" +
+                "userName='" + userName + '\'' +
+                ", userAge=" + userAge +
+                '}';
+        }
     }
-}
